Move the devices file on disk
